Hoàn thiện Main( thêm Học lực)

+++ 12_06_Buoi_16/Quan_ly_hoc_vien/Quan_Ly_Hoc_Vien.py
@@ -6,8 +6,16 @@ dulieu = ketnoi.cursor()
 def getalldata():
     dulieu.execute("SELECT * FROM quan_ly_hoc_vien.hocvien")
     ketqua = dulieu.fetchall()
+    print('{:<4} {:<15} {:<7} {:<10} {:<10} {:<13} {:<10}'.format(
+            'Id','Name','Age','Class','English','Information','Xếp loại'))
     for i in ketqua:
-        print(i)
+        diemTB = (i[-1] + i[-2])/2
+        if diemTB>5:
+            hocLuc = 'Giỏi'
+        else:
+            hocLuc = 'Khá'
+        print('{:<4} {:<15} {:<7} {:<10} {:<10} {:<13} {:<10}'.format(
+            i[0],i[1],i[2],i[3],i[4],i[5],hocLuc))
         
        
 def getalldata2():
@@ -16,6 +24,11 @@ def getalldata2():
     while ketqua is not None:
         print(ketqua)
         ketqua = dulieu.fetchone()
+        
+def getalldatatest():
+    dulieu.execute("SELECT * FROM quan_ly_hoc_vien.hocvien")
+    ketqua = dulieu.fetchall()
+    print(ketqua)
         
         
 def getdatabyid(t):
@@ -34,26 +47,49 @@ def getdatabyid2(id):
 def themHocVien():
     t = int(input('Nhập số lượng học viên muốn thêm: '))
     for i in range(0,t):
+        id =input('Id: ')
         name = input('Name: ')
         age = int(input('Age: '))
         country = input('Country: ')
         english = float(input('English: '))
         information = float(input('Information: '))
-        dulieu.execute("INSERT INTO quan_ly_hoc_vien.hocvien(`Name`,Age,Country,English,Information) VALUES('{}',{},'{}',{},{})"
-                       .format(name,age,country,english,information))
+        dulieu.execute("INSERT INTO quan_ly_hoc_vien.hocvien(Id,`Name`,Age,Country,English,Information) VALUES({},'{}',{},'{}',{},{})"
+                       .format(id,name,age,country,english,information))
         ketnoi.commit()
     
 def suaHocVien(id):
-    name = input('Name: ')
-    age = int(input('Age: '))
-    country = input('Country: ')
-    english = float(input('English: '))
-    information = float(input('Information: '))
-    sql ="UPDATE quan_ly_hoc_vien.hocvien SET `Name` = %s,Age = %s,Country = %s,English = %s,Information = %s WHERE Id = %s"
-    dulieu.execute(sql,(name,age,country,english,information,id))
-    ketnoi.commit()
+    dulieu.execute("SELECT * FROM quan_ly_hoc_vien.hocvien")
+    ketqua = dulieu.fetchall()
+    exist = False
+    for i in ketqua:
+        if i[0] == id:
+            exist = True
+            break
+    if exist == True:
+        name = input('Name: ')
+        age = int(input('Age: '))
+        country = input('Country: ')
+        english = float(input('English: '))
+        information = float(input('Information: '))
+        sql ="UPDATE quan_ly_hoc_vien.hocvien SET `Name` = %s,Age = %s,Country = %s,English = %s,Information = %s WHERE Id = %s"
+        dulieu.execute(sql,(name,age,country,english,information,id))
+        ketnoi.commit()
+    else:
+        print('Id không tồn tại!')
     
 def xoaHocVien(id):
-    sql = "DELETE FROM quan_ly_hoc_vien.hocvien WHERE Id = {}".format(id)
-    dulieu.execute(sql)
-    ketnoi.commit()
+    dulieu.execute("SELECT * FROM quan_ly_hoc_vien.hocvien")
+    ketqua = dulieu.fetchall()
+    exist = False
+    for i in ketqua:
+        if i[0] == id:
+            exist = True
+            break
+    if exist == True:
+        sql = "DELETE FROM quan_ly_hoc_vien.hocvien WHERE Id = {}".format(id)
+        dulieu.execute(sql)
+        ketnoi.commit()
+    else:
+        print('Id không tồn tại!')
+    
+#getalldatatest()
